feat: 식물 일지 이미지 연결 및 date 기반 조회 API 추가

+++ backend/src/main/java/com/plog/backend/domain/image/repository/PlantDiaryImageRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 @Repository
 public interface PlantDiaryImageRepository extends JpaRepository<PlantDiaryImage, Long> {
-    List<PlantDiaryImage> findByPlantDiaryIdAndImageIsDeletedFalseOrderByOrderAsc(int plantDiaryId);
+    List<PlantDiaryImage> findByPlantDiaryIdAndImageIsDeletedFalseOrderByOrderAsc(Long plantDiaryId);
 
-    Optional<PlantDiaryImage> findByPlantDiaryIdAndIsThumbnailTrue(int plantDiaryId);
+    Optional<PlantDiaryImage> findByPlantDiaryIdAndIsThumbnailTrue(Long plantDiaryId);
 }
